+++ corehq/warehouse/loaders/__init__.py
@@ -1,4 +1,5 @@
 from corehq.warehouse.loaders.domain import DomainStagingLoader, DomainDimLoader
+from corehq.warehouse.loaders.form import FormStagingLoader, FormFactLoader
 from corehq.warehouse.loaders.group import GroupStagingLoader, GroupDimLoader
 from corehq.warehouse.loaders.location import LocationStagingLoader, LocationDimLoader
 from corehq.warehouse.loaders.user import UserStagingLoader, UserDimLoader
@@ -14,5 +15,8 @@ def get_loader_by_slug(slug):
         GroupDimLoader,
         LocationStagingLoader,
         LocationDimLoader,
+
+        FormStagingLoader,
+        FormFactLoader,
     ]
     return {cls.slug: cls for cls in loaders}[slug]
